factor out decoration code from DB.Transactions.Model
