+++ lib/wallet/cardano-wallet.cabal
@@ -236,6 +236,7 @@ library
     Cardano.Wallet.DB.Store.Submissions.Model
     Cardano.Wallet.DB.Store.Submissions.New.Operations
     Cardano.Wallet.DB.Store.Submissions.Store
+    Cardano.Wallet.DB.Store.Transactions.Decoration
     Cardano.Wallet.DB.Store.Transactions.Layer
     Cardano.Wallet.DB.Store.Transactions.Model
     Cardano.Wallet.DB.Store.Transactions.Store
